+++ src/Permissions.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace Swe\SpaceSDK;
+
+use GuzzleHttp\Exception\GuzzleException;
+use Swe\SpaceSDK\Exception\MissingArgumentException;
+
+/**
+ * Class Permissions
+ *
+ * @package Swe\SpaceSDK
+ * @author Luca Braun <l.braun@s-w-e.com>
+ */
+class Permissions extends AbstractApi
+{
+    /**
+     * @param array $data
+     * @return bool
+     * @throws MissingArgumentException
+     * @throws GuzzleException
+     */
+    public function checkPermission(array $data): bool
+    {
+        $uri = 'permissions/check-permission';
+        $required = [
+            'principal' => 'string',
+            'uniqueRightCode' => 'string',
+            'target' => [
+                'className' => 'string',
+            ],
+        ];
+        $this->throwIfInvalid($required, $data);
+
+        return (bool)$this->client->post($this->buildUrl($uri), $data)[0];
+    }
+
+    /**
+     * @param array $response
+     * @return array
+     * @throws GuzzleException
+     */
+    public function getAllPermissions(array $response): array
+    {
+        $uri = 'permissions';
+
+        return $this->client->get($this->buildUrl($uri), $response);
+    }
+}