+++ changelog.md
@@ -29,6 +29,7 @@
 - Added class `HttpApiModel`
 - Added class `ToDoItems`
 - Added class `Blog`
+- Added class `Permissions`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`